bots/transcribinator/pyproject.toml
[project]
name = "transcribinator"
version = "1.0.0"
description = "Audio/video transcription bot using local Whisper with optional AI cleanup"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "signalinator-core",
    "requests>=2.31.0",
    "openai-whisper>=20231117",
]

[project.scripts]
transcribinator = "transcribinator.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transcribinator"]
